+++ index.html
@@ -233,6 +233,8 @@
   </template>
 
 <script src="/js/main.js"></script>
+<script src="/js/util.js"></script>
+<script src="/js/data.js"></script>
 
 </body>
 </html>
